requirements.txt
absl-py==0.13.0
asn1crypto==0.24.0
astroid==2.6.6
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cryptography==2.1.4
dataclasses==0.8
decorator==5.0.9
executing==0.8.0
flake8==3.9.2
flake8-polyfill==1.0.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
icecream==2.1.1
idna==2.6
importlib-metadata==4.5.0
importlib-resources==5.4.0
iniconfig==1.1.1
invoke==1.6.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.6.0
mando==0.6.4
Markdown==3.3.4
mccabe==0.6.1
more-itertools==8.8.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycrypto==2.6.1
pyflakes==2.3.1
Pygments==2.9.0
pygobject==3.26.1
pylint==2.9.6
pyparsing==2.4.7
pytest==6.2.4
python-apt==1.6.5+ubuntu0.6
pyxdg==0.25
PyYAML==5.4.1
radon==4.5.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
SecretStorage==2.3.1
six==1.15.0
smmap==5.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.2
termcolor==1.1.0
toml==0.10.2
tqdm==4.64.0
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.5
visual-logging==1.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xenon==0.7.3
zipp==3.4.1
